[Maps] [ES document source] Make filterByMapBounds and tooltipProperties configurations editable (#27811)
* move create source editor into its own file

* move tooltip and extent filter inputs to update editor
