+++ x-pack/plugins/gis/public/shared/layers/sources/es_search_source/index.js
@@ -1,0 +1,7 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { ESSearchSource } from './es_search_source';
